requirements.txt
alembic==1.8.0
arabic-reshaper==2.1.3
asgiref==3.4.1
aspose-words==22.7.0
atomicwrites==1.4.0
attrs==21.4.0
Authlib==0.15.5
autopep8==1.6.0
Babel==2.10.1
backports.entry-points-selectable==1.1.1
bcrypt==3.2.2
beautifulsoup4==4.11.1
bidict==0.22.0
blinker==1.4
boto3==1.20.54
botocore==1.23.54
Brotli==1.0.9
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
cryptography==3.4.8
decouple==0.0.7
defusedxml==0.7.1
distlib==0.3.3
dj-database-url==1.0.0
Django==3.2.9
django-ckeditor==6.2.0
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-scheduler==0.9.5
django-storages==1.12.3
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
django-xhtml2pdf==0.0.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dnspython==2.1.0
email-validator==1.1.3
eventlet==0.33.0
filelock==3.4.0
flake8==4.0.1
flake8-isort==4.1.1
Flask==2.0.2
Flask-Babel==2.0.0
Flask-Compress==1.12
Flask-Gravatar==0.5.0
Flask-Login==0.6.1
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-mongoengine==1.0.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.4
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fpdf==1.7.2
future==0.18.2
google-api-core==2.5.0
google-api-python-client==2.38.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
googleapis-common-protos==1.55.0
greenlet==1.1.2
gunicorn==20.1.0
html5lib==1.1
httpagentparser==1.9.2
httplib2==0.20.4
icalendar==4.0.9
idna==2.10
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
ldap3==2.9.1
lxml==4.9.1
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
mccabe==0.6.1
mongoengine==0.23.1
mysql-connector-python==8.0.28
numpy==1.22.2
oauthlib==3.1.1
packaging==15.3
paramiko==2.11.0
passlib==1.7.4
pgadmin4==6.10
Pillow==9.0.1
pipenv==2021.11.15
platformdirs==2.4.0
pluggy==1.0.0
postgres==4.0
protobuf==3.19.4
psutil==5.9.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
psycopg2-pool==1.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.10.0
PyJWT==2.3.0
pymongo==3.12.1
PyNaCl==1.5.0
pyotp==2.6.0
pyparsing==3.0.7
PyPDF2==1.26.0
pytest==7.0.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
python-docx==0.8.11
python-engineio==4.3.2
python-socketio==5.6.0
python3-openid==3.2.0
pytz==2021.3
pywinpty==1.1.6
qrcode==7.3.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.8
s3transfer==0.5.2
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.3.2.post1
speaklater3==1.4
SQLAlchemy==1.4.27
sqlparse==0.4.2
sshtunnel==0.4.0
stripe==2.63.0
testfixtures==6.18.3
toml==0.10.2
tomli==2.0.1
ua-parser==0.10.0
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.7
user-agents==2.2.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
webencodings==0.5.1
Werkzeug==2.0.3
whitenoise==6.0.0
WTForms==3.0.0
xhtml2pdf==0.2.5
